Some refacotring, and progress on Lexer, and language design.
Changed s/uword to s/uw, started to work on an array container. starting with darrays. Don't like that I don't have direct acess to length of array with zpl so going to make a custom variant of theirs.

Translating the old c code I did for the tbf engine. Made the zpl array container using the current state of the langauge design wise.
I'll stick to most likely just getting layer 1 done. I wont do layer 0 as it requires custom stuff c doesn't have and layer 0.OS doesn't have allocate and deallocate which I want in the prototype.

I will not offer any director support: alias, meta, tt.
So there will be no generics in the prototype, and no type checking.

Unless I find an easy way to do so with the parser or vistor, I'll be most likely forcing a sector attribution order for the symbols.
This would keep it c-like and unforutnately lose the flexibility of the sector but would get a boostrap done faster.

Plans are to just get lexer completely done next, by just lexging each option and then getting the parser done for a artfiical ordered set.